+++ app/controllers/timesheet_controller.rb
@@ -2,7 +2,12 @@ class TimesheetController < ApplicationController
   def show
     year = params[:year].to_i
     week_number = params[:week_number].to_i
+    @year_week_string = "#{year}-W#{params[:week_number]}"
 
+    # Stop user from accessing timesheets from before 2021
+    redirect_to('/timesheet/2021/1') if year < 2021
+
+    # Collection of dates in 'yyyy-mm-dd' format
     weekdays = []
 
     for i in 1..5 do
@@ -14,7 +19,8 @@ class TimesheetController < ApplicationController
     @entry_types = ['AM Start', 'AM Finish', 'PM Start', 'PM Finish']
     @days_of_the_week = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday']
     @week_commencing = weekdays.first
-
+    
+    # Collection of entries for each day
     @weekdays = [] 
 
     for i in 0..4 do