+++ spec/features/view_timesheet_spec.rb
@@ -38,4 +38,24 @@ feature 'view timesheets' do
     click_button('Back')
     expect(page).to have_current_path('/entries')
   end
+
+  scenario 'The week commencing date is shown on the page' do
+    visit('/timesheet/2021/46')
+    expect(page).to have_content('Week Commencing: 15/11/21')
+  end
+
+  scenario 'The week number is shown in the week form' do
+    visit('/timesheet/2021/46')
+    week_input = find('#week')
+    expect(week_input.value).to eq('2021-W46')
+
+    visit('/timesheet/2021/42')
+    week_input = find('#week')
+    expect(week_input.value).to eq('2021-W42')
+  end
+
+  scenario 'App redirects to start of 2021 if timesheet prior to 2021 accessed through url' do
+    visit('/timesheet/2020/1')
+    expect(page).to have_current_path('/timesheet/2021/1')
+  end
 end